refactor(api): migrate GET /api/campaign to prescription Bounded Context
