+++ api/src/prescription/campaign/application/campaign-detail-route.js
@@ -11,6 +11,16 @@ const register = async function (server) {
   server.route([
     {
       method: 'GET',
+      path: '/api/campaigns',
+      config: {
+        auth: false,
+        handler: campaignDetailController.getByCode,
+        notes: ['- Récupération de la campagne dont le code est spécifié dans les filtres de la requête'],
+        tags: ['api', 'campaign'],
+      },
+    },
+    {
+      method: 'GET',
       path: '/api/campaigns/{id}',
       config: {
         pre: [{ method: securityPreHandlers.checkAuthorizationToAccessCampaign }],
