+++ api/src/prescription/campaign/infrastructure/serializers/jsonapi/campaign-to-join-serializer.js
