+++ api/tests/prescription/campaign/unit/application/campaign-detail-controller_test.js
@@ -3,6 +3,54 @@ import { usecases } from '../../../../../src/prescription/campaign/domain/usecas
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Application | Controller | Campaign detail', function () {
+  describe('#getByCode', function () {
+    it('should return the serialized campaign', async function () {
+      // given
+      const code = 'AZERTY123';
+      const campaignToJoin = domainBuilder.buildCampaignToJoin({ code, identityProvider: 'SUPER_IDP' });
+      const request = {
+        query: { 'filter[code]': code },
+      };
+      sinon.stub(usecases, 'getCampaignByCode').withArgs({ code }).resolves(campaignToJoin);
+
+      // when
+      const response = await campaignDetailController.getByCode(request, hFake);
+
+      // then
+      expect(response.data).to.deep.equal({
+        type: 'campaigns',
+        id: campaignToJoin.id.toString(),
+        attributes: {
+          code: campaignToJoin.code,
+          title: campaignToJoin.title,
+          type: campaignToJoin.type,
+          'id-pix-label': campaignToJoin.idPixLabel,
+          'custom-landing-page-text': campaignToJoin.customLandingPageText,
+          'external-id-help-image-url': campaignToJoin.externalIdHelpImageUrl,
+          'alternative-text-to-external-id-help-image': campaignToJoin.alternativeTextToExternalIdHelpImage,
+          'is-archived': campaignToJoin.isArchived,
+          'is-restricted': campaignToJoin.isRestricted,
+          'is-simplified-access': campaignToJoin.isSimplifiedAccess,
+          'is-for-absolute-novice': campaignToJoin.isForAbsoluteNovice,
+          'identity-provider': campaignToJoin.identityProvider,
+          'organization-id': campaignToJoin.organizationId,
+          'organization-name': campaignToJoin.organizationName,
+          'organization-type': campaignToJoin.organizationType,
+          'organization-logo-url': campaignToJoin.organizationLogoUrl,
+          'organization-show-nps': campaignToJoin.organizationShowNPS,
+          'organization-form-nps-url': campaignToJoin.organizationFormNPSUrl,
+          'target-profile-name': campaignToJoin.targetProfileName,
+          'target-profile-image-url': campaignToJoin.targetProfileImageUrl,
+          'custom-result-page-text': campaignToJoin.customResultPageText,
+          'custom-result-page-button-text': campaignToJoin.customResultPageButtonText,
+          'custom-result-page-button-url': campaignToJoin.customResultPageButtonUrl,
+          'multiple-sendings': campaignToJoin.multipleSendings,
+          'is-flash': campaignToJoin.isFlash,
+        },
+      });
+    });
+  });
+
   describe('#getById', function () {
     const campaignId = 1;
     const userId = 1;
