+++ api/tests/prescription/campaign/unit/application/campaign-detail-route_test.js
@@ -1,9 +1,41 @@
+import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { campaignDetailController } from '../../../../../src/prescription/campaign/application/campaign-detail-controller.js';
 import * as moduleUnderTest from '../../../../../src/prescription/campaign/application/campaign-detail-route.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Application | Router | campaign-detail-router ', function () {
+  describe('GET /api/campaigns?filter[code=SOMECODE]', function () {
+    it('should return 200', async function () {
+      // given
+      sinon.stub(campaignDetailController, 'getByCode').callsFake((request, h) => h.response('ok').code(200));
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // given
+      campaignDetailController.getByCode.returns('ok');
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/campaigns?filter[code=SOMECODE]');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should return 404 when controller throws a NotFound domain error', async function () {
+      // given
+      sinon.stub(campaignDetailController, 'getByCode').rejects(new NotFoundError());
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/campaigns?filter[code=SOMECODE]');
+
+      // then
+      expect(response.statusCode).to.equal(404);
+    });
+  });
+
   describe('GET /api/campaigns/{id}', function () {
     it('should return 200', async function () {
       // given
