+++ api/tests/prescription/campaign/unit/domain/read-models/CampaignToJoin_test.js
@@ -1,5 +1,5 @@
-import { CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
-import { domainBuilder, expect } from '../../../test-helper.js';
+import { CampaignTypes } from '../../../../../../src/prescription/shared/domain/constants.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | CampaignToJoin', function () {
   describe('#isAssessment', function () {
