+++ api/tests/prescription/campaign/unit/domain/usecases/get-campaign-by-code_test.js
@@ -1,5 +1,5 @@
-import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { expect, sinon } from '../../../test-helper.js';
+import { usecases } from '../../../../../../src/prescription/campaign/domain/usecases/index.js';
+import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | get-campaign-by-code', function () {
   const code = 'QWERTY123';
