+++ api/tests/unit/application/campaign/index_test.js
@@ -1,41 +1,9 @@
 import { campaignController } from '../../../../lib/application/campaigns/campaign-controller.js';
 import { campaignStatsController } from '../../../../lib/application/campaigns/campaign-stats-controller.js';
 import * as moduleUnderTest from '../../../../lib/application/campaigns/index.js';
-import { NotFoundError } from '../../../../lib/domain/errors.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
 
 describe('Unit | Application | Router | campaign-router ', function () {
-  describe('GET /api/campaigns?filter[code=SOMECODE]', function () {
-    it('should return 200', async function () {
-      // given
-      sinon.stub(campaignController, 'getByCode').callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // given
-      campaignController.getByCode.returns('ok');
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns?filter[code=SOMECODE]');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return 404 when controller throws a NotFound domain error', async function () {
-      // given
-      sinon.stub(campaignController, 'getByCode').rejects(new NotFoundError());
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns?filter[code=SOMECODE]');
-
-      // then
-      expect(response.statusCode).to.equal(404);
-    });
-  });
-
   describe('GET /api/campaigns/{id}/collective-results', function () {
     it('should return 200', async function () {
       // given
